+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased 3.x](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/compare/3.5...HEAD)
 ### Features
+- Add routing support to SDK request types ([#351](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/pull/351))
 ### Enhancements
 ### Bug Fixes
 ### Infrastructure